requirements.txt
# Core PySpark environment
pyspark==3.5.1
findspark

# AWS SDKs for S3 interaction
boto3
botocore

# Data manipulation and validation
pandas
pyarrow
numpy
jsonschema

# Configuration and logging
python-dotenv
pyyaml
loguru

# Testing and utilities
pytest
pytest-mock

# Core data processing
pandas>=2.1.0
numpy>=1.24.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.12.2
folium>=0.15.0

# Environment & config
python-dotenv>=1.0.0

# Notebook utilities
ipython>=8.10.0
jupyter>=1.0.0
